Updated all Logout for all 3 Headers
Previously if you logged out from 1 user and logged in to another user, the homepage would be the homepage of the previous usertype, and dbUser and currentUser will be the previous user. Now, after logging out, both dbUser n currentUser will be cleared, which will ensure there is no crossover of user information between log in and out.